Django-Project/demo/requirements.txt
Django==2.2.1
django-crispy-forms==1.8.1
django-mysql==3.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
mysql-connector-python==8.0.18
mysqlclient==1.4.6
Pillow==7.0.0
protobuf==3.11.2
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
